case matters in script
The case matters in the script and I had FAT instead of Fat in there
four times. It meant that the conversion from USDA MUFA/PUFA/SFA wasn’t
being read correctly when the those things weren’t available from the
label and the script had to use the label to usda relationship for total
fat to impute a value.